// packages/contexts/ingestion/src/types.ts
export interface RateLimitConfig {
  burstLimit?: number;
  requestsPerHour?: number;
  requestsPerMinute?: number;
  requestsPerSecond: number;
}